Change name of test and functions. Move data class to main folder of test for feature login tests.
